refactor override, after and noop to a utils file
