+++ src/utils.js
@@ -1,0 +1,20 @@
+// Mostly because function props shouldn't throw errors by default
+export const noop = () => undefined;
+
+// Apply a functions that'll run after the command's function runs
+// Monkey patching for the commands
+// http://me.dt.in.th/page/JavaScript-override/
+export function override(object, methodName, callback) {
+  const dupe = object;
+  dupe[methodName] = callback(object[methodName]);
+}
+
+export function after(extraBehavior) {
+  return function(original, ...args) {
+    return function() {
+      const returnValue = original.apply(this, args);
+      extraBehavior.apply(this, args);
+      return returnValue;
+    };
+  };
+}
